+++ lib/StabilizedVehicle/src/MotorMixerNull.h
@@ -8,5 +8,5 @@ public:
 public:
     virtual void outputToMotors(const commands_t& commands, float deltaT, uint32_t tickCount) override;
     virtual float getMotorOutput(size_t motorIndex) const override;
-    virtual float getMotorRPM(size_t motorIndex) const override;
+    virtual int32_t getMotorRPM(size_t motorIndex) const override;
 };
